requirements.txt
torch
torchvision
transformers
sentence-transformers
Pillow
pandas
scikit-learn
tqdm
faiss-cpu